Refactor iqama client functions
