--- iqama.go
@@ -1,5 +1,0 @@
-package main
-
-func iqamaPrettify(b []byte) string {
-	return string(b)
-}
